+++ resources/views/goodsInWarehouse/detail.blade.php
@@ -11,7 +11,6 @@
     <small>Preview</small>
   </h1>
   <ol class="breadcrumb">
-    <li><a href="/"><i class="fa fa-dashboard"></i> Dashboard</a></li>
     <li><a href="/barang_masuk/gudang">Barang Masuk Dari Gudang</a></li>
     <li class="active">Detail Barang Masuk Dari Gudang</li>
   </ol>
@@ -21,9 +20,9 @@
 
   <div class="box box-default">
     <div class="box-header with-border">
-      <a href="/barang_masuk/gudang/detail/tambah/{{$giw->id}}" class="btn btn-md btn-info"><span class="fa fa-plus"></span></a>
-      <a href="/barang_masuk/gudang/ubah/{{$giw->id}}" class="btn btn-md btn-warning"><span class="fa fa-edit"></span></a>
-      <a href="/barang_masuk/gudang/hapus/{{$giw->id}}" class="btn btn-md btn-danger"><span class="fa fa-trash"></span></a>
+      <span data-toggle="tooltip" title="Tambah Barang"><a href="/barang_masuk/gudang/detail/tambah/{{$giw->id}}" class="btn btn-md btn-info"><span class="fa fa-plus"></span></a></span>
+      <span data-toggle="tooltip" title="Ubah Data"><a href="/barang_masuk/gudang/ubah/{{$giw->id}}" class="btn btn-md btn-warning"><span class="fa fa-edit"></span></a></span>
+      <span data-toggle="tooltip" title="Hapus Data"><a onclick="return confirm('Anda Yakin ?')" href="/barang_masuk/gudang/hapus/{{$giw->id}}" class="btn btn-md btn-danger"><span class="fa fa-trash"></span></a></span>
 
       <div class="box-tools pull-right">
         <button type="button" class="btn btn-box-tool" data-widget="collapse"><i class="fa fa-minus"></i></button>
@@ -70,7 +69,7 @@
       </div>
     </div>
     <div class="box-body">
-      <table class="table table-hovered table-bordered">
+      <table id="table" class="table table-hovered table-bordered">
         <thead>
           <tr>
             <th width="1%"><center>No</center></th>
@@ -94,8 +93,8 @@
             <td>{{$item->description}}</td>
             <td>
               <center>
-                <a href="/barang_masuk/gudang/detail/ubah/{{$item->id}}" class="btn btn-md btn-warning"><span class="fa fa-edit"></span></a>
-                <a href="/barang_masuk/gudang/detail/hapus/{{$item->id}}" class="btn btn-md btn-danger"><span class="fa fa-trash"></span></a>
+                <span data-toggle="tooltip" title="Ubah Data"><a href="/barang_masuk/gudang/detail/ubah/{{$item->id}}" class="btn btn-md btn-warning"><span class="fa fa-edit"></span></a></span>
+                <span data-toggle="tooltip" title="Hapus Data"><a onclick="return confirm('Anda Yakin ?')" href="/barang_masuk/gudang/detail/hapus/{{$item->id}}" class="btn btn-md btn-danger"><span class="fa fa-trash"></span></a></span>
               </center>
             </td>
           </tr>
@@ -107,4 +106,17 @@
   @endif
 
 </section>
+@endsection
+
+@section('js')
+<script type="text/javascript">
+  $(document).ready(function(){
+    $("#table").DataTable({
+      "columnDefs": [
+        { "orderable": false, "targets": 0},
+        { "orderable": false, "targets": 5},
+      ]
+    });
+  });
+</script>
 @endsection
